+++ backend/src/main/java/com/debede/backend/repository/CarritoRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CarritoRepository extends JpaRepository<Carrito, Long> {
+public interface CarritoRepository extends JpaRepository<Carrito, Integer> {
 
 }
